% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NaSub}
\alias{NaSub}
\title{Replaces NAs with either the last non-NA observation, or a specified replacement value}
\usage{
NaSub(x, repl = 0, last_obs = TRUE, maxgap = 0L)
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{repl}{The value used to replace NAs}

\item{last_obs}{TRUE if the last non-NA observation should be used instead of a specified replacement value}

\item{maxgap}{Runs of more than \code{maxgap} are retained. By default, the argument is set to 0, which 
will preserve all NAs. If set to -1, the argument is set to infinity such that sequences of any size of NAs
will be replaced with the last non-NA observation.}
}
\description{
Replaces NAs with either the last non-NA observation, or a specified replacement value
}

