% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingBeta}
\alias{RollingBeta}
\title{Estimates the coefficient (beta) from a univariate linear model y ~ x over a rolling window}
\usage{
RollingBeta(x, y, window, pop = FALSE, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix; the independent variable}

\item{y}{A vector or single-column matrix; the dependent variable}

\item{window}{The size of the rolling window}

\item{pop}{TRUE to calculate population (n) covariance, FALSE to calculate sample (n-1) covariance}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Estimates \eqn{\hat{\beta_1}} from a ordinary least squares model \eqn{y = \beta_0 + \beta_1 x}.
Beta is defined as
\deqn{ \code{Beta} = \frac{cov(x, y)}{\sigma^2_x} }
where \eqn{cov} is covariance and \eqn{\sigma^2} is variance.
}

