% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingCorr}
\alias{RollingCorr}
\title{Calculates Pearson correlation over a rolling window}
\usage{
RollingCorr(x, y, window, pop = FALSE, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix}

\item{y}{A vector or single-column matrix}

\item{window}{The size of the rolling window}

\item{pop}{TRUE to calculate population (n) std. deviation, FALSE to calculate sample (n-1) std. deviation}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Correlation is defined as
\deqn{ \code{Corr} = \frac{cov(x, y)}{\sigma_x\sigma_y} }
where \eqn{cov} is the covariance and \eqn{\sigma} is the standard deviation
}

