% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingMAE}
\alias{RollingMAE}
\title{Calculates the mean absolute error over a rolling window}
\usage{
RollingMAE(x, y, window, expanding = FALSE, na_method = "none")
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{y}{A vector, matrix, list or other object coercible to a vector}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}

\item{na_method}{One of "none" (default), "ignore" or "window" (see notes on NA handling)}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Calculates the mean absolute error over a rolling window
}
\details{
Mean absolute error is defined as
\deqn{ \code{MAE} = \frac{1}{n}\sum^n_{i=1}|x_i - y_i| }
where \eqn{x} and \eqn{y} represent the predicted and actual (true) values.
}

