% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingMedian}
\alias{RollingMedian}
\title{Calculates the median over a rolling window}
\usage{
RollingMedian(x, window, expanding = FALSE, na_method = "none")
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}

\item{na_method}{One of "none" (default), "ignore" or "window" (see notes on NA handling)}
}
\value{
A matrix; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Calculates the median over a rolling window
}
\details{
If input \code{x} contains more than one column, the rolling window calculation will be performed on each
column. The dimensions of the return value will be the same as input \code{x}.
}

