% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingSharpe}
\alias{RollingSharpe}
\title{Calculates the Sharpe Ratio over a rolling window}
\usage{
RollingSharpe(x, Rf, window, expanding = FALSE, na_method = "none",
  scale = 12)
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{Rf}{The risk-free rate, which is must be a vector, matrix, list or other object coercible to a vector.
Rf must have the same number of observations as x.}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}

\item{na_method}{One of "none" (default), "ignore" or "window" (see notes on NA handling)}

\item{scale}{Used in the compounding exponent: product^(scale / window) - 1; any number (integer or floating point) may be used}
}
\value{
A matrix; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Calculates the Sharpe Ratio over a rolling window
}
\details{
The Sharpe Ratio is defined as
\deqn{ \code{Zscore} = \frac{x - Rf}{\sigma} }
where \eqn{\sigma} is the standard deviation of x.
}

