% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingSkew}
\alias{RollingSkew}
\title{Calculates the skewness over a rolling window}
\usage{
RollingSkew(x, window, expanding = FALSE, na_method = "none", pop = FALSE)
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}

\item{na_method}{One of "none" (default), "ignore" or "window" (see notes on NA handling)}

\item{pop}{TRUE to calculate population (n) statistic, FALSE to calculate sample (n-1) statistic}
}
\value{
A matrix; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Calculates the skewness over a rolling window
}
\details{
Population skewness (see parameter \code{pop}) is defined as
\deqn{ \code{Skew} = \sqrt{n}\frac{\sum^n_{i=1}(x_i-\bar{x})^3}{\big(\sum^n_{i=1}(x_i-\bar{x})^2\big)^{3/2}}}
Sample skewness (see parameter \code{pop}) is defined as
\deqn{ \code{Skew} = \frac{n\sqrt{n-1}}{n-2}\frac{\sum^n_{i=1}(x_i-\bar{x})^3}{\big(\sum^n_{i=1}(x_i-\bar{x})^2\big)^{3/2}}}
}

