% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingSS}
\alias{RollingSS}
\title{Calculates the sum of squares over a rolling window}
\usage{
RollingSS(x, window, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Sum of squares is defined as
\deqn{ \code{SS} = \sum^n_{i=1} x^2_i }
}

