% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingSSE}
\alias{RollingSSE}
\title{Calculates the sum of squared errors over a rolling window}
\usage{
RollingSSE(x, y, window, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix}

\item{y}{A vector or single-column matrix; the error to be subtracted from x}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Sum of squared errors is defined as
\deqn{ \code{SSE} = \sum^n_{i=1}(x_i - y_i)^2 }
where \eqn{x} and \eqn{y} represent the predicted and actual (true) values.
}

