% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingSumprod}
\alias{RollingSumprod}
\title{Calculates the sum product (dot product) over a rolling window}
\usage{
RollingSumprod(x, y, window, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix}

\item{y}{A vector or single-column matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A vector; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Sum product is defined as
\deqn{ \code{Sumprod} = \sum^n_{i=1} x_i y_i }
}

