% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingMean}
\alias{RollingMean}
\title{Calculates the arithmetic mean over a rolling window}
\usage{
RollingMean(x, window, expanding = FALSE)
}
\arguments{
\item{x}{A vector or single-column matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}
}
\value{
A matrix; rows 1 to (length(window) - 1) contain NAs.
}
\description{
The arithmetic mean is defined as
\deqn{\code{Mean} = \frac{1}{n}\sum^n_{i=1} x_i}
}
\details{
If input \code{x} contains more than one column, the rolling window calculation will be performed on each
column. The dimensions of the return value will be the same as input \code{x}.
}

