% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RollingStd}
\alias{RollingStd}
\title{Calculates the standard deviation over a rolling window}
\usage{
RollingStd(x, window, expanding = FALSE, na_rm = FALSE, pop = FALSE)
}
\arguments{
\item{x}{A vector, matrix, list or other object coercible to a matrix}

\item{window}{The size of the rolling window}

\item{expanding}{TRUE to calculate an expanding window instead of a rolling window; if TRUE, window
determines the starting number of observations used to calculate the statistic}

\item{na_rm}{TRUE to handle NAs (see notes on NA handling)}

\item{pop}{TRUE to calculate population (n) statistic, FALSE to calculate sample (n-1) statistic}
}
\value{
A matrix; rows 1 to (length(window) - 1) contain NAs.
}
\description{
Calculates the standard deviation over a rolling window
}
\details{
Sample standard deviation (see parameter \code{pop}) is defined as
\deqn{ \code{Std} = \sqrt{\frac{1}{n-1}\sum^n_{i=1}(x_i - \bar{x})^2} }
}

